#161 Componentizando pagina e Resolvendo problema no carrousel
